+++ hazelcast-extension/src/main/java/org/jnosql/artemis/hazelcast/kv/Param.java
@@ -12,7 +12,7 @@
  *
  *   Otavio Santana
  */
-package org.jnosql.artemis.hazelcast.key;
+package org.jnosql.artemis.hazelcast.kv;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
