+++ hazelcast-extension/src/test/java/org/jnosql/artemis/hazelcast/kv/DefaultHazelcastTemplateTest.java
@@ -12,7 +12,7 @@
  *
  *   Otavio Santana
  */
-package org.jnosql.artemis.hazelcast.key;
+package org.jnosql.artemis.hazelcast.kv;
 
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
