+++ hazelcast-extension/src/test/java/org/jnosql/artemis/hazelcast/kv/Person.java
@@ -12,7 +12,7 @@
  *
  *   Otavio Santana
  */
-package org.jnosql.artemis.hazelcast.key;
+package org.jnosql.artemis.hazelcast.kv;
 
 
 import jakarta.nosql.mapping.Column;
